requirements.txt
Pillow>=9.0.0
pyserial>=3.5 